% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChkRetVals_2.R
\name{ChkRetVals_2_F1}
\alias{ChkRetVals_2_F1}
\title{ChkRetVals_2 F1}
\usage{
ChkRetVals_2_F1(a, c)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}
}
\value{
xy
\itemize{
   \item x
   \item y
}
}
\description{
ChkRetVals_2 F1
}
