% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChkRetVals_2.R
\name{ChkRetVals_2_F4}
\alias{ChkRetVals_2_F4}
\title{ChkRetVals_2 F4}
\usage{
ChkRetVals_2_F4(a, b)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}
}
\value{
xy
\itemize {
   \item x
   \item y
}
}
\description{
ChkRetVals_2 F4
}
