% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cycle_2.R
\name{Cycle_1_Func1}
\alias{Cycle_1_Func1}
\title{Cycle_2 Func1}
\usage{
Cycle_1_Func1(a, b)
}
\arguments{
\item{x}{parameter x}

\item{y}{parameter y}
}
\value{
xy
\itemize{
   \item x
   \item y
}
}
\description{
Cycle_2 Func1
}
