% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_Heli_pine.R, R/u_Heli_spruce.R
\name{Density}
\alias{Density}
\title{u_Heli_pine Density}
\usage{
Density(dh, ch, ts, heli = 1.11, a1 = 304.3, b1 = 10.4437,
  b2 = 444.13, b3 = 0.2957)

Density(dh, ch, ts, heli = 1.11, a1 = 304.3, b1 = 10.4437,
  b2 = 444.13, b3 = 0.2957)
}
\arguments{
\item{ts}{temperature sum}

\item{diameter_h}{diameter under bark at height h}

\item{age_h}{number of annual rings (age) at height h}

\item{ageb_h}{number of annual rings at breast height (1.3 m)}

\item{diameter_bh}{diameter under bark at breast height}

\item{Heli_a}{Heli_a term}

\item{diameter_h}{diameter under bark at height h}

\item{age_h}{number of annual rings (age) at height h}

\item{tempsum}{temperature sum}

\item{Heli_d}{Heli_d term}
}
\value{
density

density
}
\description{
u_Heli_pine Density

u_Heli_spruce Density
}
