% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_Heli_pine.R, R/u_Heli_spruce.R
\name{Earlywood}
\alias{Earlywood}
\title{u_Heli_pine Earlywood}
\usage{
Earlywood(lw, heli = 0.95)

Earlywood(lw, heli = 0.95)
}
\arguments{
\item{latewood}{latewood}

\item{Heli_b}{Heli_b term}

\item{latewood}{latewood}

\item{Heli_e}{Heli_e term}
}
\value{
earlywood

earlywood
}
\description{
u_Heli_pine Earlywood

u_Heli_spruce Earlywood
}
