% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eohakkid_kuusi}
\alias{Eohakkid_kuusi}
\title{funcHakkila Eohakkid_kuusi}
\usage{
Eohakkid_kuusi(dd, a1 = -3.71, b1 = 0.10229, b2 = 3.3e-06)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBM
}
\description{
funcHakkila Eohakkid_kuusi
}
