% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eohakkid_manty}
\alias{Eohakkid_manty}
\title{funcHakkila Eohakkid_manty}
\usage{
Eohakkid_manty(dd, a1 = -9.3954, b1 = 2.3268)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBM
}
\description{
funcHakkila Eohakkid_manty
}
