% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eohakkidh_kuusi}
\alias{Eohakkidh_kuusi}
\title{funcHakkila Eohakkidh_kuusi}
\usage{
Eohakkidh_kuusi(dd, hh, a1 = 0.4112, b1 = 0.00026724, b2 = 1.41e-06,
  b3 = 0.00043562)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\description{
funcHakkila Eohakkidh_kuusi
}
