% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eohakkidh_manty}
\alias{Eohakkidh_manty}
\title{funcHakkila Eohakkidh_manty}
\usage{
Eohakkidh_manty(dd, hh, a1 = -5.2678, b2 = 3.4914, b3 = 1.9498,
  b4 = 47.454)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\description{
funcHakkila Eohakkidh_manty
}
