% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eohakkidhcr_kuusi}
\alias{Eohakkidhcr_kuusi}
\title{funcHakkila Eohakkidhcr_kuusi}
\usage{
Eohakkidhcr_kuusi(dd, hh, ss, a1 = -9.821, b1 = 2.3031,
  b2 = 0.017075, b3 = 992.36)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}

\item{height}{height of the tree}
}
\description{
funcHakkila Eohakkidhcr_kuusi
}
