% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eohakkidhcr_manty}
\alias{Eohakkidhcr_manty}
\title{funcHakkila Eohakkidhcr_manty}
\usage{
Eohakkidhcr_manty(dd, hh, ss, a1 = -11.059, b1 = 2.5605,
  b2 = 0.0058932, b3 = 8.14e-05, b4 = 1.13e-06)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}

\item{height}{height of the tree}
}
\description{
funcHakkila Eohakkidhcr_manty
}
