% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcKorhonen_Maltamo.R
\name{Eokorhodhcr}
\alias{Eokorhodhcr}
\title{funcKorhonen_Maltamo Eokorhodhcr}
\usage{
Eokorhodhcr(dd, hh, ss, a1 = -3.0599, b1 = 3.1287, b2 = 1.3936,
  b3 = 0.0005298, b4 = 0.1162)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{crownratio}{crownratio of the tree}
}
\value{
LBM
}
\description{
funcKorhonen_Maltamo Eokorhodhcr
}
