% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Eomarkd_kuusi}
\alias{Eomarkd_kuusi}
\title{funcMarklund Eomarkd_kuusi}
\usage{
Eomarkd_kuusi(dd, a1 = 1.2804, b1 = 8.5242)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBM
}
\description{
funcMarklund Eomarkd_kuusi
}
