% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Eomarkd_manty}
\alias{Eomarkd_manty}
\title{funcMarklund Eomarkd_manty}
\usage{
Eomarkd_manty(dd, a1 = 2.8604, b1 = 9.1015)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBM
}
\description{
funcMarklund Eomarkd_manty
}
