% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Eomarkdh_kuusi}
\alias{Eomarkdh_kuusi}
\title{funcMarklund Eomarkdh_kuusi}
\usage{
Eomarkdh_kuusi(dd, hh, a1 = 1.2063, b1 = 10.9708, b2 = 0.0124,
  b3 = 0.4923)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBM
}
\description{
funcMarklund Eomarkdh_kuusi
}
