% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Eomarkdh_manty}
\alias{Eomarkdh_manty}
\title{funcMarklund Eomarkdh_manty}
\usage{
Eomarkdh_manty(dd, hh, a1 = 2.5413, b1 = 13.3955, b2 = 1.1955)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBM
}
\description{
funcMarklund Eomarkdh_manty
}
