% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Eonfiner1}
\alias{Eonfiner1}
\title{funcFiner Eonfiner1}
\usage{
Eonfiner1(dd, a1 = 3.2988, b1 = 2.2387)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBNM
}
\description{
funcFiner Eonfiner1
}
