% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Eonfiner2_koivu}
\alias{Eonfiner2_koivu}
\title{funcFiner Eonfiner2_koivu}
\usage{
Eonfiner2_koivu(dd, a1 = 3.3891)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBNM
}
\description{
funcFiner Eonfiner2_koivu
}
