% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Eonfiner2_manty}
\alias{Eonfiner2_manty}
\title{funcFiner Eonfiner2_manty}
\usage{
Eonfiner2_manty(dd, a1 = 3.056)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBNM
}
\description{
funcFiner Eonfiner2_manty
}
