% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Eonfiner3}
\alias{Eonfiner3}
\title{funcFiner Eonfiner3}
\usage{
Eonfiner3(dd, ss, a1 = 2.7932, b1 = 2.6708, b2 = 1.3568)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}
}
\value{
LBNM
}
\description{
funcFiner Eonfiner3
}
