% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Eonfiner4}
\alias{Eonfiner4}
\title{funcFiner Eonfiner4}
\usage{
Eonfiner4(dd, ss, a1 = 4.1857, b1 = 2.20822, b2 = 0.6252)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}
}
\value{
LBNM
}
\description{
funcFiner Eonfiner4
}
