% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eonhakkid}
\alias{Eonhakkid}
\title{funcHakkila Eonhakkid}
\usage{
Eonhakkid(dd, a1 = -10.7699, b1 = 2.6016)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcHakkila Eonhakkid
}
