% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Eonhakkidhcr}
\alias{Eonhakkidhcr}
\title{funcHakkila Eonhakkidhcr}
\usage{
Eonhakkidhcr(dd, hh, ss, a1 = -12.4606, b1 = 2.73067, b2 = 1788.9,
  b3 = 0.01664)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}

\item{height}{height of the tree}
}
\value{
LBNM
}
\description{
funcHakkila Eonhakkidhcr
}
