% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Eonlaiho_koivu}
\alias{Eonlaiho_koivu}
\title{funcLaiho Eonlaiho_koivu}
\usage{
Eonlaiho_koivu(dd, hh, hd, b1 = 3.775, b2 = 2.966, b3 = -0.745)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
LBNM
}
\description{
funcLaiho Eonlaiho_koivu
}
