% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Eonlaiho_manty}
\alias{Eonlaiho_manty}
\title{funcLaiho Eonlaiho_manty}
\usage{
Eonlaiho_manty(dd, hh, ss, b1 = 3.992, b2 = 3.285, b3 = -0.804)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
LBNM
}
\description{
funcLaiho Eonlaiho_manty
}
