% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Eonmarkd}
\alias{Eonmarkd}
\title{funcMarklund Eonmarkd}
\usage{
Eonmarkd(dd, a1 = 3.3633, b1 = 10.2806)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
LBMN
}
\description{
funcMarklund Eonmarkd
}
