% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malli_1.R, R/malli_1_ParRaj.R
\name{F0}
\alias{F0}
\title{malli_1 F0}
\usage{
F0(a, c)

F0(a, c)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}

\item{a}{parameter a}

\item{c}{parameter c}
}
\value{
x

x
}
\description{
malli_1 F0

malli_1_ParRaj F0
}
