% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malli_1.R, R/malli_1_ParRaj.R
\name{F1}
\alias{F1}
\title{malli_1 F1}
\usage{
F1(a, c)

F1(a, c)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}

\item{b}{parameter b}
}
\value{
c

x
}
\description{
malli_1 F1

malli_1_ParRaj F1
}
