% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malli_1.R, R/malli_1_ParRaj.R
\name{F2}
\alias{F2}
\title{malli_1 F2}
\usage{
F2(a, c)

F2(a, c)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}

\item{a}{parameter a}
}
\value{
c

x
}
\description{
malli_1 F2

malli_1_ParRaj F2
}
