% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_Heli_pine.R, R/u_Heli_spruce.R
\name{Fibre}
\alias{Fibre}
\title{u_Heli_pine Fibre}
\usage{
Fibre(dh, ch, ts, rh, heli = 1.05, a1 = 3.06, b1 = 2.2, b2 = 29.35,
  b3 = 1.55, b4 = -0.5, b5 = 0.74, b6 = 0.7, b7 = 0.14,
  b8 = 0.00056)

Fibre(dh, ch, ts, rh, heli = 1.05, a1 = 3.06, b1 = 2.2, b2 = 29.35,
  b3 = 1.55, b4 = -0.5, b5 = 0.74, b6 = 0.7, b7 = 0.14,
  b8 = 0.00056)
}
\arguments{
\item{heli}{Heli_f}

\item{diameter_h}{diameter under bark at height h}

\item{age_h}{number of annual rings (age) at height h}

\item{tempsum}{temperature sum}

\item{relative_h}{relative height}

\item{Heli_c}{Heli_c term}

\item{diameter_h}{diameter under bark at height h}

\item{age_h}{number of annual rings (age) at height h}

\item{tempsum}{temperature sum}

\item{relative_h}{relative height of the tree}
}
\value{
fibre_l

fibre_l
}
\description{
u_Heli_pine Fibre

u_Heli_spruce Fibre
}
