% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Hukka}
\alias{Hukka}
\title{funcR_osat Hukka}
\usage{
Hukka(BM, vv, kk, vw, kw)
}
\arguments{
\item{frametimbermass}{runkopuun (kuoreton) massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{barkvolume}{volume of the bark}

\item{unbarkedvolumewastewood}{hukkapuun kuorellinen tilavuus}

\item{wastebarkvolume}{hukkapuun kuoren tilavuus}
}
\value{
WBM kuorettoman hukkapuun massa
}
\description{
funcR_osat Hukka
}
