% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Jmarkd_koivu2}
\alias{Jmarkd_koivu2}
\title{funcMarklund Jmarkd_koivu2}
\usage{
Jmarkd_koivu2(dd, a1 = 6.3413, a2 = 3.8375, b1 = 13.2902,
  b2 = 8.8795)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
RM
}
\description{
funcMarklund Jmarkd_koivu2
}
