% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Jmarkd_kuusi1}
\alias{Jmarkd_kuusi1}
\title{funcMarklund Jmarkd_kuusi1}
\usage{
Jmarkd_kuusi1(dd, b1 = 7.6283, b2 = 2.5706)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
RM
}
\description{
funcMarklund Jmarkd_kuusi1
}
