% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Jmarkd_kuusi2}
\alias{Jmarkd_kuusi2}
\title{funcMarklund Jmarkd_kuusi2}
\usage{
Jmarkd_kuusi2(dd, a1 = 6.3851, a2 = 2.5706, b1 = 13.3703,
  b2 = 7.6283)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
RM
}
\description{
funcMarklund Jmarkd_kuusi2
}
