% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Jmarkd_manty1}
\alias{Jmarkd_manty1}
\title{funcMarklund Jmarkd_manty1}
\usage{
Jmarkd_manty1(dd, a1 = 3.8375, b1 = 8.8795)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
RM
}
\description{
funcMarklund Jmarkd_manty1
}
