% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kamarkd_kuusi}
\alias{Kamarkd_kuusi}
\title{funcMarklund Kamarkd_kuusi}
\usage{
Kamarkd_kuusi(dd, a1 = 3.3645, b1 = 10.6686)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STM
}
\description{
funcMarklund Kamarkd_kuusi
}
