% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kamarkd_muut}
\alias{Kamarkd_muut}
\title{funcMarklund Kamarkd_muut}
\usage{
Kamarkd_muut(dd, a1 = 3.9657, b1 = 11.0481)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STM
}
\description{
funcMarklund Kamarkd_muut
}
