% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keohakkid_kuusi}
\alias{Keohakkid_kuusi}
\title{funcHakkila Keohakkid_kuusi}
\usage{
Keohakkid_kuusi(dd, a1 = -4.34, b1 = 0.11571, b2 = 3.34e-06)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\description{
funcHakkila Keohakkid_kuusi
}
