% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keohakkid_manty}
\alias{Keohakkid_manty}
\title{funcHakkila Keohakkid_manty}
\usage{
Keohakkid_manty(dd, a1 = -8.8027, b1 = 2.2475)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\description{
funcHakkila Keohakkid_manty
}
