% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keohakkidh_kuusi}
\alias{Keohakkidh_kuusi}
\title{funcHakkila Keohakkidh_kuusi}
\usage{
Keohakkidh_kuusi(dd, hh, a1 = 0.4866, b1 = 0.00035026, b2 = 1.35e-06,
  b3 = 0.00042424)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\description{
funcHakkila Keohakkidh_kuusi
}
