% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keohakkidh_manty}
\alias{Keohakkidh_manty}
\title{funcHakkila Keohakkidh_manty}
\usage{
Keohakkidh_manty(dd, hh, a1 = -9.7486, b1 = 0.0016023, b2 = 2.56,
  b3 = 0.0063173)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
DLBM
}
\description{
funcHakkila Keohakkidh_manty
}
