% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keohakkidhcr_kuusi}
\alias{Keohakkidhcr_kuusi}
\title{funcHakkila Keohakkidhcr_kuusi}
\usage{
Keohakkidhcr_kuusi(dd, hh, ss, a1 = -9.3308, b1 = 2.2742,
  b2 = 872.51, b3 = 0.013594)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}

\item{height}{height of the tree}
}
\description{
funcHakkila Keohakkidhcr_kuusi
}
