% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keohakkidhcr_manty}
\alias{Keohakkidhcr_manty}
\title{funcHakkila Keohakkidhcr_manty}
\usage{
Keohakkidhcr_manty(dd, hh, ss, a1 = -10.23, b1 = 2.5478,
  b2 = 0.0050869, b3 = 0.0057087)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}

\item{height}{height of the tree}
}
\value{
DLBM
}
\description{
funcHakkila Keohakkidhcr_manty
}
