% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keonhakkid}
\alias{Keonhakkid}
\title{funcHakkila Keonhakkid}
\usage{
Keonhakkid(dd, a1 = -10.2692, b1 = 2.5124)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DLNM
}
\description{
funcHakkila Keonhakkid
}
