% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Keonhakkidhcr}
\alias{Keonhakkidhcr}
\title{funcHakkila Keonhakkidhcr}
\usage{
Keonhakkidhcr(dd, ss, a1 = -9.0521, b1 = 2.0532, b2 = 5.498e-05)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}
}
\description{
funcHakkila Keonhakkidhcr
}
