% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kfiner1}
\alias{Kfiner1}
\title{funcFiner Kfiner1}
\usage{
Kfiner1(dd, a1 = -4.0168, b1 = 2.0023)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcFiner Kfiner1
}
