% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kfiner2_koivu}
\alias{Kfiner2_koivu}
\title{funcFiner Kfiner2_koivu}
\usage{
Kfiner2_koivu(dd, a1 = -5.5507, b1 = 2.7731)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcFiner Kfiner2_koivu
}
