% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kfiner2_manty}
\alias{Kfiner2_manty}
\title{funcFiner Kfiner2_manty}
\usage{
Kfiner2_manty(dd, a1 = -2.9847, b1 = 1.6874)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcFiner Kfiner2_manty
}
