% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kfiner3}
\alias{Kfiner3}
\title{funcFiner Kfiner3}
\usage{
Kfiner3(dd, hh, a1 = -0.722, b1 = 2.1274, b2 = 0.7392)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BAM
}
\description{
funcFiner Kfiner3
}
