% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kfiner4}
\alias{Kfiner4}
\title{funcFiner Kfiner4}
\usage{
Kfiner4(dd, a1 = -5.912, b1 = 2.7644)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcFiner Kfiner4
}
