% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Khukka}
\alias{Khukka}
\title{funcR_osat Khukka}
\usage{
Khukka(BAM, kk, kw)
}
\arguments{
\item{trunkbarkmass}{rungon kuoren massa}

\item{barkvolume}{volume of the bark}

\item{wastebarkvolume}{hukkapuun kuoren tilavuus}
}
\value{
BRWM mass of the bark in wastewood
}
\description{
funcR_osat Khukka
}
