% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kjfiner1_koivu}
\alias{Kjfiner1_koivu}
\title{funcFiner Kjfiner1_koivu}
\usage{
Kjfiner1_koivu(dd, a = -5.3806, b1 = 3.0861)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcFiner Kjfiner1_koivu
}
