% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kjfiner1_kuusi}
\alias{Kjfiner1_kuusi}
\title{funcFiner Kjfiner1_kuusi}
\usage{
Kjfiner1_kuusi(dd, a1 = -4.9853, b1 = 3.0333)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcFiner Kjfiner1_kuusi
}
