% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kjfiner1_manty}
\alias{Kjfiner1_manty}
\title{funcFiner Kjfiner1_manty}
\usage{
Kjfiner1_manty(dd, a = -4.5698, b1 = 2.7929)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcFiner Kjfiner1_manty
}
