% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{KJfiner2}
\alias{KJfiner2}
\title{funcFiner KJfiner2}
\usage{
KJfiner2(dd, a1 = -3.342, b1 = 2.4142)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcFiner KJfiner2
}
