% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Kjhakkid_kuusi}
\alias{Kjhakkid_kuusi}
\title{funcHakkila Kjhakkid_kuusi}
\usage{
Kjhakkid_kuusi(dd, a1 = -7, b1 = 0.051, b2 = 1.33)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcHakkila Kjhakkid_kuusi
}
