% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Kjhakkid_muut}
\alias{Kjhakkid_muut}
\title{funcHakkila Kjhakkid_muut}
\usage{
Kjhakkid_muut(dd, a1 = -2, b1 = 0.039, b2 = 1.33)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\description{
funcHakkila Kjhakkid_muut
}
