% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kjmarkd_kuusi}
\alias{Kjmarkd_kuusi}
\title{funcMarklund Kjmarkd_kuusi}
\usage{
Kjmarkd_kuusi(dd, a1 = 2.4447, b1 = 10.5381)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcMarklund Kjmarkd_kuusi
}
