% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kjmarkd_muut}
\alias{Kjmarkd_muut}
\title{funcMarklund Kjmarkd_muut}
\usage{
Kjmarkd_muut(dd, a1 = 3.3913, b1 = 11.1106)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
STRM
}
\description{
funcMarklund Kjmarkd_muut
}
