% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcKorhonen_Maltamo.R
\name{Kkorhodh}
\alias{Kkorhodh}
\title{funcKorhonen_Maltamo Kkorhodh}
\usage{
Kkorhodh(dd, hh, a1 = -4.344, b1 = 0.885, b2 = 0.435)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BAM
}
\description{
funcKorhonen_Maltamo Kkorhodh
}
