% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcKorhonen_Maltamo.R
\name{Kkorhodhbo_st5}
\alias{Kkorhodhbo_st5}
\title{funcKorhonen_Maltamo Kkorhodhbo_st5}
\usage{
Kkorhodhbo_st5(dd, hh, a1 = -4.409, b1 = 0.905, b2 = 0.3925,
  b3 = 0.101)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BAM
}
\description{
funcKorhonen_Maltamo Kkorhodhbo_st5
}
