% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Klaiho_koivu}
\alias{Klaiho_koivu}
\title{funcLaiho Klaiho_koivu}
\usage{
Klaiho_koivu(dd, hh, b1 = 5.156, b2 = 1.197, b3 = 1.456)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BAM
}
\description{
funcLaiho Klaiho_koivu
}
