% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Klaiho_manty}
\alias{Klaiho_manty}
\title{funcLaiho Klaiho_manty}
\usage{
Klaiho_manty(dd, hh, b1 = 5.658, b2 = 2.251, b3 = 0.249)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BAM
}
\description{
funcLaiho Klaiho_manty
}
