% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Klatvak}
\alias{Klatvak}
\title{funcR_osat Klatvak}
\usage{
Klatvak(BAM, kk, kt)
}
\arguments{
\item{topstemfiberbarkmass}{rungon kuoren massa}

\item{barkvolume}{volume of the bark}

\item{barkvolumetopfibre}{latvakuitupuun kuoren tilavuus}
}
\value{
BRTM latvakuitupuun kuoren massa
}
\description{
funcR_osat Klatvak
}
