% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kmarkd_koivu}
\alias{Kmarkd_koivu}
\title{funcMarklund Kmarkd_koivu}
\usage{
Kmarkd_koivu(dd, a1 = 3.2518, b1 = 10.3876)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcMarklund Kmarkd_koivu
}
