% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kmarkd_kuusi}
\alias{Kmarkd_kuusi}
\title{funcMarklund Kmarkd_kuusi}
\usage{
Kmarkd_kuusi(dd, a1 = 3.3912, b1 = 9.8364)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcMarklund Kmarkd_kuusi
}
