% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kmarkd_manty}
\alias{Kmarkd_manty}
\title{funcMarklund Kmarkd_manty}
\usage{
Kmarkd_manty(dd, a1 = 2.9748, b1 = 8.8489)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcMarklund Kmarkd_manty
}
