% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kmarkdh_koivu}
\alias{Kmarkdh_koivu}
\title{funcMarklund Kmarkdh_koivu}
\usage{
Kmarkdh_koivu(dd, hh, a1 = 8.3019, b1 = 0.7433, b2 = 4.0778)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcMarklund Kmarkdh_koivu
}
