% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kmarkdh_kuusi}
\alias{Kmarkdh_kuusi}
\title{funcMarklund Kmarkdh_kuusi}
\usage{
Kmarkdh_kuusi(dd, hh, a1 = 3.402, b1 = 8.3089, b2 = 0.0147,
  b3 = 0.2295)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcMarklund Kmarkdh_kuusi
}
