% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Kmarkdh_manty}
\alias{Kmarkdh_manty}
\title{funcMarklund Kmarkdh_manty}
\usage{
Kmarkdh_manty(dd, hh, a1 = 0.4487, b1 = 7.2482, b2 = 3.2765)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BAM
}
\description{
funcMarklund Kmarkdh_manty
}
