% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kofiner1}
\alias{Kofiner1}
\title{funcFiner Kofiner1}
\usage{
Kofiner1(dd, a = 2.9225, b1 = 1.7067)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcFiner Kofiner1
}
