% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kofiner2_koivu}
\alias{Kofiner2_koivu}
\title{funcFiner Kofiner2_koivu}
\usage{
Kofiner2_koivu(dd, ss, a = -19.7244, b1 = 5.929, b2 = 10.9833)
}
\arguments{
\item{crownratio}{crownratio of the tree}

\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcFiner Kofiner2_koivu
}
