% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kofiner2_manty}
\alias{Kofiner2_manty}
\title{funcFiner Kofiner2_manty}
\usage{
Kofiner2_manty(dd, a = 2.7189)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcFiner Kofiner2_manty
}
