% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kofiner3}
\alias{Kofiner3}
\title{funcFiner Kofiner3}
\usage{
Kofiner3(dd, a = 1.8943, b1 = 2.2181)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcFiner Kofiner3
}
