% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Kofiner4}
\alias{Kofiner4}
\title{funcFiner Kofiner4}
\usage{
Kofiner4(dd, a = 5.5372, b1 = 0.1328)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcFiner Kofiner4
}
