% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Kohakkid_koivu}
\alias{Kohakkid_koivu}
\title{funcHakkila Kohakkid_koivu}
\usage{
Kohakkid_koivu(dd, a1 = -0.07, b1 = 0.004)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\description{
funcHakkila Kohakkid_koivu
}
