% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Kohakkid_kuusi}
\alias{Kohakkid_kuusi}
\title{funcHakkila Kohakkid_kuusi}
\usage{
Kohakkid_kuusi(dd, a1 = -0.62, b1 = 0.0134, b2 = 3.9e-08)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\description{
funcHakkila Kohakkid_kuusi
}
