% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Kohakkid_manty}
\alias{Kohakkid_manty}
\title{funcHakkila Kohakkid_manty}
\usage{
Kohakkid_manty(dd, a1 = -0.84, b1 = 0.0194)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcHakkila Kohakkid_manty
}
