% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Kolaiho_koivu}
\alias{Kolaiho_koivu}
\title{funcLaiho Kolaiho_koivu}
\usage{
Kolaiho_koivu(dd, hh, b1 = 1130.121, b2 = 1.693, b3 = -2.146)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
DBM
}
\description{
funcLaiho Kolaiho_koivu
}
