% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Kolaiho_manty}
\alias{Kolaiho_manty}
\title{funcLaiho Kolaiho_manty}
\usage{
Kolaiho_manty(dd, hh, hd, b1 = 3101.458, b2 = 0.307, b3 = 5.275)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
DBM
}
\description{
funcLaiho Kolaiho_manty
}
