% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkd_koivu}
\alias{Komarkd_koivu}
\title{funcMarklund Komarkd_koivu}
\usage{
Komarkd_koivu(dd, a1 = 5.9507, b1 = 7.9266)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkd_koivu
}
