% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkd_kuusi}
\alias{Komarkd_kuusi}
\title{funcMarklund Komarkd_kuusi}
\usage{
Komarkd_kuusi(dd, a1 = 4.3308, b1 = 9.955)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkd_kuusi
}
