% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkd_manty}
\alias{Komarkd_manty}
\title{funcMarklund Komarkd_manty}
\usage{
Komarkd_manty(dd, a1 = 5.3338, b1 = 9.5938)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkd_manty
}
