% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkdh_koivu}
\alias{Komarkdh_koivu}
\title{funcMarklund Komarkdh_koivu}
\usage{
Komarkdh_koivu(dd, hh, a1 = 6.6237, b1 = 11.2872, b2 = 0.3081,
  b3 = 2.6821)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkdh_koivu
}
