% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkdh_kuusi}
\alias{Komarkdh_kuusi}
\title{funcMarklund Komarkdh_kuusi}
\usage{
Komarkdh_kuusi(dd, hh, a1 = 4.6351, b1 = 3.6518, b2 = 0.0493,
  b3 = 1.0129)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkdh_kuusi
}
