% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkdh_manty}
\alias{Komarkdh_manty}
\title{funcMarklund Komarkdh_manty}
\usage{
Komarkdh_manty(dd, hh, a1 = 5.8926, b1 = 7.127, b2 = 0.0465,
  b3 = 1.106)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkdh_manty
}
