% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Komarkdhcr}
\alias{Komarkdhcr}
\title{funcMarklund Komarkdhcr}
\usage{
Komarkdhcr(dd, hh, ss, a1 = 5.3924, b1 = 5.6333, b2 = 2.7826,
  b3 = 1.746)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
DBM
}
\description{
funcMarklund Komarkdhcr
}
