% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Krhakki1_koivu}
\alias{Krhakki1_koivu}
\title{funcHakkila Krhakki1_koivu}
\usage{
Krhakki1_koivu(dd, tt, vv, a1 = 379.99, b1 = 22.84, b2 = 2.771,
  b3 = 1.01, b4 = 1.005)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}
}
\description{
funcHakkila Krhakki1_koivu
}
