% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Krhakki1_kuusi2}
\alias{Krhakki1_kuusi2}
\title{funcHakkila Krhakki1_kuusi2}
\usage{
Krhakki1_kuusi2(dd, hh, tt, vv, a1 = 307.21, b1 = 67.35, b2 = 0.27,
  b3 = 0.001679, b4 = 167.7, b5 = 9.837, b6 = 17.79, b7 = 1.01)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}
}
\value{
BABM
}
\description{
funcHakkila Krhakki1_kuusi2
}
