% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Krhakki1_manty2}
\alias{Krhakki1_manty2}
\title{funcHakkila Krhakki1_manty2}
\usage{
Krhakki1_manty2(dd, hh, tt, vv, a1 = 381.63, b1 = 27.43, b2 = 62.44,
  b3 = 26.88, b4 = 0.526, b5 = 1.01, b6 = 0.98, b7 = 0.97)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}
}
\value{
BABM
}
\description{
funcHakkila Krhakki1_manty2
}
