% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Krhakki2_koivu}
\alias{Krhakki2_koivu}
\title{funcHakkila Krhakki2_koivu}
\usage{
Krhakki2_koivu(dd, tt, vv, a1 = 335.64, b1 = 34.156, b2 = 138.5,
  b3 = 1.01, b4 = 1.005)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}
}
\description{
funcHakkila Krhakki2_koivu
}
