% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Krhakki2_kuusi2}
\alias{Krhakki2_kuusi2}
\title{funcHakkila Krhakki2_kuusi2}
\usage{
Krhakki2_kuusi2(dd, hh, tt, vv, a1 = 303.37, b1 = 67.95, b2 = 0.2795,
  b3 = 0.619, b4 = 19.13, b5 = 1.01)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}
}
\description{
funcHakkila Krhakki2_kuusi2
}
