% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Krhakki2_manty2}
\alias{Krhakki2_manty2}
\title{funcHakkila Krhakki2_manty2}
\usage{
Krhakki2_manty2(dd, hh, tt, vv, a1 = 341.77, b1 = 92.93, b2 = 193,
  b3 = 1.832, b4 = 1.01, b5 = 0.98, b6 = 0.97)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}
}
\description{
funcHakkila Krhakki2_manty2
}
