% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Krhukka}
\alias{Krhukka}
\title{funcR_osat Krhukka}
\usage{
Krhukka(BABM, vv, vw)
}
\arguments{
\item{unbarkedmass}{kuorellisen rungon massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{unbarkedvolumewastewood}{hukkapuun kuorellinen tilavuu}
}
\value{
BATM kuorellisen latvakuitupuun massa
}
\description{
funcR_osat Krhukka
}
