% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Krlaiho}
\alias{Krlaiho}
\title{funcLaiho Krlaiho}
\usage{
Krlaiho(dd, hh, b1 = 38.711, b2 = 1.789, b3 = 0.901)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BABM
}
\description{
funcLaiho Krlaiho
}
