% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Krlatvak}
\alias{Krlatvak}
\title{funcR_osat Krlatvak}
\usage{
Krlatvak(BABM, vv, vt)
}
\arguments{
\item{unbarkedmass}{kuorellisen rungon massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{unbarkedvolumetopfiber}{latvakuitupuun kuorellinen tilavuus}
}
\value{
BATM kuorellisen latvakuitupuun massa
}
\description{
funcR_osat Krlatvak
}
