% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkd_koivu}
\alias{Krmarkd_koivu}
\title{funcMarklund Krmarkd_koivu}
\usage{
Krmarkd_koivu(dd, a1 = 3.0932, b1 = 11.0735)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkd_koivu
}
