% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkd_kuusi}
\alias{Krmarkd_kuusi}
\title{funcMarklund Krmarkd_kuusi}
\usage{
Krmarkd_kuusi(dd, a1 = 2.0571, b1 = 11.3341)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkd_kuusi
}
