% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkd_manty}
\alias{Krmarkd_manty}
\title{funcMarklund Krmarkd_manty}
\usage{
Krmarkd_manty(dd, a1 = 2.3388, b1 = 11.3264)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkd_manty
}
