% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkdh_koivu}
\alias{Krmarkdh_koivu}
\title{funcMarklund Krmarkdh_koivu}
\usage{
Krmarkdh_koivu(dd, hh, a1 = 3.5686, b1 = 8.2827, b2 = 0.0393,
  b3 = 0.5772)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkdh_koivu
}
