% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkdh_kuusi}
\alias{Krmarkdh_kuusi}
\title{funcMarklund Krmarkdh_kuusi}
\usage{
Krmarkdh_kuusi(dd, hh, a1 = 2.1702, b1 = 7.469, b2 = 0.0289,
  b3 = 0.6828)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkdh_kuusi
}
