% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkdh_manty}
\alias{Krmarkdh_manty}
\title{funcMarklund Krmarkdh_manty}
\usage{
Krmarkdh_manty(dd, hh, a1 = 2.6768, b1 = 7.5939, b2 = 0.0151,
  b3 = 0.8799)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkdh_manty
}
