% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Krmarkdht}
\alias{Krmarkdht}
\title{funcMarklund Krmarkdht}
\usage{
Krmarkdht(dd, hh, ad, a1 = 3.5194, b1 = 8.042, b2 = 0.0531,
  b3 = 0.3897, b4 = 0.1018)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BABM
}
\description{
funcMarklund Krmarkdht
}
