% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Krrunkok}
\alias{Krrunkok}
\title{funcR_osat Krrunkok}
\usage{
Krrunkok(BABM, vv, vp)
}
\arguments{
\item{unbarkedmass}{kuorellisen rungon massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{unbarkedvolumetrunkfiber}{runkokuitupuun kuorellinen tilavuus}
}
\value{
BAPM kuorellinen runkokuitupuun massa
}
\description{
funcR_osat Krrunkok
}
