% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Krtukki}
\alias{Krtukki}
\title{funcR_osat Krtukki}
\usage{
Krtukki(BABM, vv, vs)
}
\arguments{
\item{unbarkedmass}{kuorellisen rungon massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{unbarkedvolumetimber}{tukkipuun kuorellinen tilavuus}
}
\value{
BASM kuorellisen tukkipuun massa
}
\description{
funcR_osat Krtukki
}
