% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Krunkok}
\alias{Krunkok}
\title{funcR_osat Klatvak}
\usage{
Krunkok(BAM, kk, kp)
}
\arguments{
\item{trunkbarkmass}{rungon kuoren massa}

\item{barkvolume}{volume of the bark}

\item{stemfiberbarkvolume}{runkokuitupuun kuoren tilavuus}
}
\value{
BRWM mass of the bark in stem fiber tree
}
\description{
funcR_osat Klatvak
}
