% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Ktukki}
\alias{Ktukki}
\title{funcR_osat Ktukki}
\usage{
Ktukki(BAM, kk, ks)
}
\arguments{
\item{trunkbarkmass}{rungon kuoren massa}

\item{barkvolume}{volume of the bark}

\item{timbervolume}{tukkipuun kuoren tilavuus}
}
\value{
BRSM tukkipuun kuoren massa
}
\description{
funcR_osat Ktukki
}
