% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_Heli_pine.R, R/u_Heli_spruce.R
\name{Latewood}
\alias{Latewood}
\title{u_Heli_pine Latewood}
\usage{
Latewood(dh, ch, ts, a1 = 6.1, b1 = 9.183, b2 = 28.885,
  b3 = 0.005911)

Latewood(dh, ch, ts, a1 = 6.1, b1 = 9.183, b2 = 28.885,
  b3 = 0.005911)
}
\arguments{
\item{ts}{temperature sum}

\item{diameter_h}{diameter under bark at height h}

\item{age_h}{number of annual rings (age) at height h}

\item{age_bh}{number of annual rings at breast height (1.3 m)}

\item{diameter_bh}{diameter under bark at breast height}

\item{Heli_a}{Heli_a term}

\item{diameter_h}{diameter under bark at height h}

\item{age_h}{number of annual rings (age) at height h}

\item{tempsum}{temperature sum}
}
\value{
latewood

latewood
}
\description{
u_Heli_pine Latewood

u_Heli_spruce Latewood
}
