% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Latvak}
\alias{Latvak}
\title{funcR_osat Latvak}
\usage{
Latvak(BM, vv, kk, vt, kt)
}
\arguments{
\item{frametimbermass}{runkopuun (kuoreton) massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{barkvolume}{volume of the bark}

\item{unbarkedvolumetopfiber}{latvakuitupuun kuorellinen tilavuus}

\item{barkvolumetopfibre}{latvakuitupuun kuoren tilavuus}
}
\value{
TBM kuorettoman latvakuitupuun massa
}
\description{
funcR_osat Latvak
}
