% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malli_1_EiParRaj.R
\name{malli_1_EiParRaj_F1}
\alias{malli_1_EiParRaj_F1}
\title{malli_1_EiParRaj F0}
\usage{
malli_1_EiParRaj_F1(a, c)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}
}
\value{
x
}
\description{
malli_1_EiParRaj F0
}
