% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malli_1_EiParRaj.R
\name{malli_1_EiParRaj_F4}
\alias{malli_1_EiParRaj_F4}
\title{malli_1_EiParRaj F0}
\usage{
malli_1_EiParRaj_F4(a, b)
}
\arguments{
\item{a}{parameter a}

\item{b}{parameter b}
}
\value{
c
}
\description{
malli_1_EiParRaj F0
}
