% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malli_1_ParRaj.R
\name{malli_1_ParRaj_F3}
\alias{malli_1_ParRaj_F3}
\title{malli_1_ParRaj F3}
\usage{
malli_1_ParRaj_F3(a, c)
}
\arguments{
\item{a}{parameter a}

\item{c}{parameter c}
}
\value{
x
}
\description{
malli_1_ParRaj F3
}
