% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcKorhonen_Maltamo.R
\name{Mpkorhodhcr}
\alias{Mpkorhodhcr}
\title{funcKorhonen_Maltamo Mpkorhodhcr}
\usage{
Mpkorhodhcr(dd, hh, ss, a1 = -2.649, b1 = 2.127, b2 = 0.117,
  b3 = 0.000202, b4 = 0.1496, b5 = 0.00163)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{crownratio}{crownratio of the tree}
}
\value{
AGM
}
\description{
funcKorhonen_Maltamo Mpkorhodhcr
}
