% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Mplaiho_koivu}
\alias{Mplaiho_koivu}
\title{funcLaiho Mplaiho_koivu}
\usage{
Mplaiho_koivu(dd, hh, b1 = 37.437, b2 = 2.031, b3 = 0.799)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
AGM
}
\description{
funcLaiho Mplaiho_koivu
}
