% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Mplaiho_kuusi}
\alias{Mplaiho_kuusi}
\title{funcLaiho Mplaiho_kuusi}
\usage{
Mplaiho_kuusi(dd, hh, b1 = 92.558, b2 = 1.652, b3 = 0.855)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
AGM
}
\description{
funcLaiho Mplaiho_kuusi
}
