% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Mplaiho_manty}
\alias{Mplaiho_manty}
\title{funcLaiho Mplaiho_manty}
\usage{
Mplaiho_manty(dd, hh, b1 = 53.677, b2 = 2.143, b3 = 0.51)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
AGM
}
\description{
funcLaiho Mplaiho_manty
}
