% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Nfiner1}
\alias{Nfiner1}
\title{funcFiner Nfiner1}
\usage{
Nfiner1(dd, a1 = 2.3958, a2 = 3.0319, a3 = 2.845, b1 = 1.8121,
  b2 = 1.666, b3 = 1.7205)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcFiner Nfiner1
}
