% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Nfiner2_koivu}
\alias{Nfiner2_koivu}
\title{funcFiner Nfiner2_koivu}
\usage{
Nfiner2_koivu(dd, b1 = 2.9749)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcFiner Nfiner2_koivu
}
