% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Nfiner2_manty}
\alias{Nfiner2_manty}
\title{funcFiner Nfiner2_manty}
\usage{
Nfiner2_manty(dd, a1 = 1.1906, a2 = 1.8862, a3 = 2.1337,
  b1 = 2.0091, b2 = 1.8809, b3 = 1.8666)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcFiner Nfiner2_manty
}
