% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Nfiner3}
\alias{Nfiner3}
\title{funcFiner Nfiner3}
\usage{
Nfiner3(dd, ss, a1 = 2.8966, a2 = 2.6514, a3 = 2.6905, b1 = 1.932,
  b2 = 1.9828, b3 = 1.8779, b4 = 1.1458, b5 = 1.152, b6 = 0.8357)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{}
}
\value{
NM
}
\description{
funcFiner Nfiner3
}
