% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Nfiner4}
\alias{Nfiner4}
\title{funcFiner Nfiner4}
\usage{
Nfiner4(dd, ss, a1 = 2.2939, a2 = 2.8349, a3 = 2.776, a4 = 2.5818,
  a5 = 2.4523, a6 = 2.1302, a7 = 1.9123, b1 = 1.8981,
  b2 = 0.8337, b3 = 1.7137, b4 = 0.5291, b5 = 1.6714,
  b6 = 0.5235, b7 = 1.782, b8 = 0.6142, b9 = 1.7122,
  b10 = 0.5289, b11 = 1.8232, b12 = 0.7358, b13 = 2.1511,
  b14 = 0.9628)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{}
}
\value{
NM
}
\description{
funcFiner Nfiner4
}
