% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Nhakkid_kuusi}
\alias{Nhakkid_kuusi}
\title{funcHakkila Nhakkid_kuusi}
\usage{
Nhakkid_kuusi(dd, a1 = -9.03, b1 = 2.2204)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcHakkila Nhakkid_kuusi
}
