% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Nhakkid_manty}
\alias{Nhakkid_manty}
\title{funcHakkila Nhakkid_manty}
\usage{
Nhakkid_manty(dd, a1 = -7.47, b1 = 1.6975)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcHakkila Nhakkid_manty
}
