% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Nhakkidcr_kuusi}
\alias{Nhakkidcr_kuusi}
\title{funcHakkila Nhakkidcr_kuusi}
\usage{
Nhakkidcr_kuusi(dd, ss, a1 = 0.37, b1 = 1.592e-08, b2 = 4.73e-06)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}
}
\value{
NM
}
\description{
funcHakkila Nhakkidcr_kuusi
}
