% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Nhakkidcr_manty}
\alias{Nhakkidcr_manty}
\title{funcHakkila Nhakkidcr_manty}
\usage{
Nhakkidcr_manty(dd, ss, a1 = -9.01, b1 = 1.8485, b2 = 0.0155)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{crownratio}{crownratio of the tree}
}
\description{
funcHakkila Nhakkidcr_manty
}
