% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Nlaiho_koivu}
\alias{Nlaiho_koivu}
\title{funcLaiho Nlaiho_koivu}
\usage{
Nlaiho_koivu(dd, b1 = 13.18, b2 = 1.978)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcLaiho Nlaiho_koivu
}
