% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Nlaiho_manty}
\alias{Nlaiho_manty}
\title{funcLaiho Nlaiho_manty}
\usage{
Nlaiho_manty(dd, hh, ss, b1 = 33.203, b2 = 2.132, b3 = -0.502)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
NM
}
\description{
funcLaiho Nlaiho_manty
}
