% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Nmarkd_kuusi}
\alias{Nmarkd_kuusi}
\title{funcMarklund Nmarkd_kuusi}
\usage{
Nmarkd_kuusi(dd, a1 = 1.9602, b1 = 7.8171)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcMarklund Nmarkd_kuusi
}
