% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Nmarkd_manty}
\alias{Nmarkd_manty}
\title{funcMarklund Nmarkd_manty}
\usage{
Nmarkd_manty(dd, a1 = 3.7983, b1 = 7.7681)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcMarklund Nmarkd_manty
}
