% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Nmarkdh_kuusi}
\alias{Nmarkdh_kuusi}
\title{funcMarklund Nmarkdh_kuusi}
\usage{
Nmarkdh_kuusi(dd, hh, a1 = 1.8551, b1 = 9.7809, b2 = 0.4873)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcMarklund Nmarkdh_kuusi
}
