% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Nmarkdh_muut}
\alias{Nmarkdh_muut}
\title{funcMarklund Nmarkdh_muut}
\usage{
Nmarkdh_muut(dd, hh, a1 = 3.4781, b1 = 12.1095, b2 = 0.0413,
  b3 = 1.565)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcMarklund Nmarkdh_muut
}
