% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Nmarkdhcr}
\alias{Nmarkdhcr}
\title{funcMarklund Nmarkdhcr}
\usage{
Nmarkdhcr(dd, hh, ss, a1 = 1.5732, b1 = 8.4127, b2 = 1.5628,
  b3 = 1.4032)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
NM
}
\description{
funcMarklund Nmarkdhcr
}
