% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_Heli.R
\name{RelativeHeight}
\alias{RelativeHeight}
\title{misc_Heli RelativeHeight}
\usage{
RelativeHeight(x, h)
}
\arguments{
\item{h}{parameter h}

\item{h_x}{paramter h_x}
}
\value{
rel_h
}
\description{
misc_Heli RelativeHeight
}
