% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Rfiner1}
\alias{Rfiner1}
\title{funcFiner Rfiner1}
\usage{
Rfiner1(dd, hh, a1 = -3.2875, b1 = 1.8136, b2 = 0.9317)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcFiner Rfiner1
}
