% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Rfiner2_koivu}
\alias{Rfiner2_koivu}
\title{funcFiner Rfiner2_koivu}
\usage{
Rfiner2_koivu(dd, a1 = -2.3362, b1 = 2.3811)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcFiner Rfiner2_koivu
}
