% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Rfiner2_manty}
\alias{Rfiner2_manty}
\title{funcFiner Rfiner2_manty}
\usage{
Rfiner2_manty(dd, a1 = -1.6475, b1 = 2.1064)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcFiner Rfiner2_manty
}
