% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Rfiner3}
\alias{Rfiner3}
\title{funcFiner Rfiner3}
\usage{
Rfiner3(dd, hh, a1 = -5.7103, b1 = 1.7256, b2 = 1.0241)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcFiner Rfiner3
}
