% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFiner.R
\name{Rfiner4}
\alias{Rfiner4}
\title{funcFiner Rfiner4}
\usage{
Rfiner4(dd, hh, a1 = -2.8414, b1 = 1.5541, b2 = 1.056)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcFiner Rfiner4
}
