% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Rhakki1_koivu}
\alias{Rhakki1_koivu}
\title{funcHakkila Rhakki1_koivu}
\usage{
Rhakki1_koivu(dd, tt, vv, kk, a1 = 379.99, b1 = 22.84, b2 = 2.771,
  b3 = 1.01)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}

\item{knotbarkcorrection}{knot-bark correction term}
}
\value{
BM
}
\description{
funcHakkila Rhakki1_koivu
}
