% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Rhakki1_kuusi1}
\alias{Rhakki1_kuusi1}
\title{funcHakkila Rhakki1_kuusi1}
\usage{
Rhakki1_kuusi1(dd, hh, tt, vv, kk, a1 = 307.21, b1 = 67.35,
  b2 = 0.27, b3 = 0.001679, b4 = 167.7, b5 = 9.837, b6 = 17.79,
  b7 = 1.01, b8 = 1.05)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}

\item{knotbarkcorrection}{knot-bark correction term}
}
\value{
BM
}
\description{
funcHakkila Rhakki1_kuusi1
}
