% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Rhakki1_manty1}
\alias{Rhakki1_manty1}
\title{funcHakkila Rhakki1_manty1}
\usage{
Rhakki1_manty1(dd, hh, tt, vv, kk, a1 = 381.63, b1 = 27.43,
  b2 = 62.44, b3 = 26.88, b4 = 0.526, b5 = 1.01, b6 = 0.93)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}

\item{knotbarkcorrection}{knot-bark correction term}
}
\value{
BM
}
\description{
funcHakkila Rhakki1_manty1
}
