% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Rhakki2_koivu}
\alias{Rhakki2_koivu}
\title{funcHakkila Rhakki2_koivu}
\usage{
Rhakki2_koivu(dd, tt, vv, kk, a1 = 335.64, b1 = 34.156, b2 = 138.5,
  b3 = 1.01)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}

\item{knotbarkcorrection}{knot-bark correction term}
}
\description{
funcHakkila Rhakki2_koivu
}
