% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Rhakki2_kuusi1}
\alias{Rhakki2_kuusi1}
\title{funcHakkila Rhakki2_kuusi1}
\usage{
Rhakki2_kuusi1(dd, hh, tt, vv, kk, a1 = 303.37, b1 = 67.95,
  b2 = 0.2795, b3 = 0.619, b4 = 19.13, b5 = 1.01, b6 = 1.05)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}

\item{knotbarkcorrection}{knot-bark correction term}
}
\description{
funcHakkila Rhakki2_kuusi1
}
