% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHakkila.R
\name{Rhakki2_manty3}
\alias{Rhakki2_manty3}
\title{funcHakkila Rhakki2_manty3}
\usage{
Rhakki2_manty3(dd, hh, tt, vv, kk, a1 = 341.77, b1 = 92.93, b2 = 193,
  b3 = 1.832, b4 = 1.01)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}

\item{age}{age of the tree}

\item{volume}{volume of the tree}

\item{knotbarkcorrection}{knot-bark correction term}
}
\value{
BM
}
\description{
funcHakkila Rhakki2_manty3
}
