% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcKorhonen_Maltamo.R
\name{Rkorhodh}
\alias{Rkorhodh}
\title{funcKorhonen_Maltamo Rkorhodh}
\usage{
Rkorhodh(dd, hh, a1 = -4.182, b1 = 0.879, b2 = 1.215)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcKorhonen_Maltamo Rkorhodh
}
