% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcKorhonen_Maltamo.R
\name{Rkorhodht}
\alias{Rkorhodht}
\title{funcKorhonen_Maltamo Rkorhodht}
\usage{
Rkorhodht(dd, hh, tt, a1 = -4.326, b1 = 0.842, b2 = 1.212,
  b3 = 0.087)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcKorhonen_Maltamo Rkorhodht
}
