% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Rlaiho_koivu}
\alias{Rlaiho_koivu}
\title{funcLaiho Rlaiho_koivu}
\usage{
Rlaiho_koivu(dd, hh, b1 = 6.329, b2 = 1.82, b3 = 1.58)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcLaiho Rlaiho_koivu
}
