% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLaiho.R
\name{Rlaiho_manty}
\alias{Rlaiho_manty}
\title{funcLaiho Rlaiho_manty}
\usage{
Rlaiho_manty(dd, hh, b1 = 14.422, b2 = 1.84, b3 = 1.185)
}
\arguments{
\item{diameter}{mean diameter of the tree}

\item{height}{height of the tree}
}
\value{
BM
}
\description{
funcLaiho Rlaiho_manty
}
