% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Rmarkd_koivu}
\alias{Rmarkd_koivu}
\title{funcMarklund Rmarkd_koivu}
\usage{
Rmarkd_koivu(dd, a1 = 2.3327, b1 = 10.8109)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcMarklund Rmarkd_koivu
}
