% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Rmarkd_kuusi}
\alias{Rmarkd_kuusi}
\title{funcMarklund Rmarkd_kuusi}
\usage{
Rmarkd_kuusi(dd, a1 = 2.2471, b1 = 11.4873)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcMarklund Rmarkd_kuusi
}
