% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Rmarkd_manty}
\alias{Rmarkd_manty}
\title{funcMarklund Rmarkd_manty}
\usage{
Rmarkd_manty(dd, a1 = 2.2184, b1 = 11.4219)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcMarklund Rmarkd_manty
}
