% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Rmarkdh_koivu}
\alias{Rmarkdh_koivu}
\title{funcMarklund Rmarkdh_koivu}
\usage{
Rmarkdh_koivu(dd, hh, a1 = 3.3045, b1 = 8.1184, b2 = 0.9783)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcMarklund Rmarkdh_koivu
}
