% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Rmarkdh_kuusi}
\alias{Rmarkdh_kuusi}
\title{funcMarklund Rmarkdh_kuusi}
\usage{
Rmarkdh_kuusi(dd, hh, a1 = 2.3032, b1 = 7.2309, b2 = 0.0355,
  b3 = 0.703)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcMarklund Rmarkdh_kuusi
}
