% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcMarklund.R
\name{Rmarkdh_manty}
\alias{Rmarkdh_manty}
\title{funcMarklund Rmarkdh_manty}
\usage{
Rmarkdh_manty(dd, hh, a1 = 2.6864, b1 = 7.6066, b2 = 0.02,
  b3 = 0.8658)
}
\arguments{
\item{diameter}{mean diameter of the tree}
}
\value{
BM
}
\description{
funcMarklund Rmarkdh_manty
}
