% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Runkok}
\alias{Runkok}
\title{funcR_osat Runkok}
\usage{
Runkok(BM, vv, kk, vp, kp)
}
\arguments{
\item{frametimbermass}{runkopuun (kuoreton) massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{barkvolume}{volume of the bark}

\item{unbarkedvolumetrunkfiber}{runkokuitupuun kuorellinen tilavuus}

\item{stemfiberbarkvolume}{runkokuitupuun kuoren tilavuus}
}
\value{
PBM kuorettoman runkokuitupuun massa
}
\description{
funcR_osat Runkok
}
