% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcR_osat.R
\name{Tukki}
\alias{Tukki}
\title{funcR_osat Tukki}
\usage{
Tukki(BM, vv, kk, vs, ks)
}
\arguments{
\item{frametimbermass}{runkopuun (kuoreton) massa}

\item{unbarkedvolume}{kuorellisen rungon tilavuus}

\item{barkvolume}{volume of the bark}

\item{unbarkedvolumetimber}{tukkipuun kuorellinen tilavuus}

\item{timbervolume}{tukkipuun kuoren tilavuus}
}
\value{
SWTM
}
\description{
funcR_osat Tukki
}
